+++ NEWS.md
@@ -1,6 +1,6 @@
 # T4transport 0.1.1
 
-* Support for special inputs such as histograms and grayscale images added.
+* Support for special types of data. See the reference for more details.
 
 # T4transport 0.1.0