+++ NEWS.md
@@ -1,3 +1,7 @@
+# T4transport 0.1.2
+
+* `gaussmed1d()` and `gaussmedpd()` now fully respects the product manifold perspective.
+
 # T4transport 0.1.1
 
 * Support for special types of data. See the reference for more details.